+++ changelog.md
@@ -4,5 +4,6 @@
 - Changed intersection algorithm
 - Fixed bug that made the player unable to jump when playing a level
 - Fixed bug that didnt hide controller glyphs or button hints when going into gameplay
+- Added "Disable "Players Destroyed" stats incrementing" and "Disable extra margin"
 ## v1.0.0
 - Initial Release!